requirements.txt
PyYAML==5.4.1
requests==2.26.0
python-decouple==3.6